refactor: refactor rata-summary checks
