+++ src/qa-certification-workspace/qa-certification-checks.service.spec.ts
@@ -12,6 +12,7 @@ import { LinearitySummaryImportDTO } from '../dto/linearity-summary.dto';
 import { LinearityInjectionImportDTO } from '../dto/linearity-injection.dto';
 import { BadRequestException } from '@nestjs/common';
 import { RataChecksService } from '../rata-workspace/rata-checks.service';
+import { RataSummaryChecksService } from '../rata-summary-workspace/rata-summary-checks.service';
 
 const returnLocationRunChecks = [
   {
@@ -59,6 +60,12 @@ describe('QA Certification Check Service Test', () => {
         },
         {
           provide: RataChecksService,
+          useFactory: () => ({
+            runChecks: jest.fn().mockResolvedValue([]),
+          }),
+        },
+        {
+          provide: RataSummaryChecksService,
           useFactory: () => ({
             runChecks: jest.fn().mockResolvedValue([]),
           }),
