+++ src/rata-run-workspace/rata-run.controller.ts
@@ -1,4 +1,12 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+} from '@nestjs/common';
 import {
   ApiCreatedResponse,
   ApiOkResponse,
@@ -81,7 +89,7 @@ export class RataRunWorkspaceController {
     const userId = 'testUser';
     return this.service.deleteRataRun(testSumId, id, userId);
   }
-  
+
   @Put(':id')
   //  @ApiBearerAuth('Token')
   //  @UseGuards(AuthGuard)
