+++ src/rata-run-workspace/rata-run.service.spec.ts
@@ -41,7 +41,6 @@ const mockRepository = () => ({
   find: jest.fn().mockResolvedValue([rataRun]),
   findOne: jest.fn().mockResolvedValue(rataRun),
   create: jest.fn().mockResolvedValue(rataRun),
-  save: jest.fn().mockResolvedValue(rataRun),
   delete: jest.fn().mockReturnValue(''),
 });
 
@@ -131,6 +130,8 @@ describe('RataRunWorkspaceService', () => {
       const result = await service.deleteRataRun(testSumId, rataRunId, userId);
       expect(repository.delete).toHaveBeenCalled();
       expect(testSummaryService.resetToNeedsEvaluation).toHaveBeenCalled();
+    });
+  });
 
   describe('updateRataRun', () => {
     it('should update a rata run record', async () => {
