+++ src/rata-workspace/rata-checks.service.spec.ts
@@ -12,6 +12,7 @@ import { RataFrequencyCodeRepository } from '../rata-frequency-code/rata-frequen
 import { TestSummaryWorkspaceRepository } from '../test-summary-workspace/test-summary.repository';
 import { RataChecksService } from './rata-checks.service';
 
+const locationId = '';
 const testSumId = '';
 
 const monitorSystemRecord = new MonitorSystem();
@@ -72,6 +73,7 @@ describe('Rata Checks Service Test', () => {
   describe('RATA Checks', () => {
     it('Should pass all checks for Import', async () => {
       const result = await service.runChecks(
+        locationId,
         importPayload,
         null,
         testSummaryImportPayload,
@@ -81,7 +83,11 @@ describe('Rata Checks Service Test', () => {
     });
 
     it('Should pass all checks for POST or UPDATE', async () => {
-      const result = await service.runChecks(importPayload, testSumId);
+      const result = await service.runChecks(
+        locationId,
+        importPayload,
+        testSumId,
+      );
       expect(result).toEqual([]);
     });
   });
@@ -97,7 +103,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-102-C] The value [${importPayload.numberOfLoadLevels}] in the field [numberOfLoadLevels] for [RATA] is not within the range of valid values.`,
@@ -109,7 +115,7 @@ describe('Rata Checks Service Test', () => {
       importPayload.numberOfLoadLevels = 0;
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-102-B] The value [${importPayload.numberOfLoadLevels}] in the field [numberOfLoadLevels] for [RATA] is not within the range of valid values from [1] to [3].`,
@@ -134,7 +140,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-103-A] You reported [relativeAccuracy], which is not appropriate for [].`,
@@ -157,7 +163,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-103-B] You did not provide [relativeAccuracy], which is required for [RATA].`,
@@ -180,7 +186,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-103-C] The value [-1] in the field [relativeAccuracy] for [RATA] is not within the range of valid values. This value must be greater than or equal to zero.`,
@@ -205,7 +211,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-104-A] You reported [overallBiasAdjustmentFactor], which is not appropriate for [${testSumRec.testTypeCode}].`,
@@ -228,7 +234,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-104-B] You did not provide [overallBiasAdjustmentFactor], which is required for [RATA].`,
@@ -251,7 +257,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-104-C] The value [${importPayload.overallBiasAdjustmentFactor}] in the field [overallBiasAdjustmentFactor] for [RATA] is not within the range of valid values. This value must be greater than or equal to 1.000.`,
@@ -276,7 +282,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-105-A] You reported [rataFrequencyCode], which is not appropriate for [${testSumRec.testTypeCode}].`,
@@ -299,7 +305,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(testSumRec);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-105-B] You did not provide [rataFrequencyCode], which is required for [RATA].`,
@@ -325,7 +331,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(null);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-105-C] You reported the value [${importPayload.rataFrequencyCode}], which is not in the list of valid values, in the field [rataFrequencyCode] for [RATA].`,
@@ -353,7 +359,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(rataFreqCdRecord);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-105-C] You reported the value [${importPayload.rataFrequencyCode}], which is not in the list of valid values, in the field [rataFrequencyCode] for [RATA].`,
@@ -382,7 +388,7 @@ describe('Rata Checks Service Test', () => {
         .mockResolvedValue(rataFreqCdRecord);
 
       try {
-        await service.runChecks(importPayload, testSumId);
+        await service.runChecks(locationId, importPayload, testSumId);
       } catch (err) {
         expect(err.response.message).toEqual([
           `[RATA-105-C] You reported the value [${importPayload.rataFrequencyCode}], which is not in the list of valid values, in the field [rataFrequencyCode] for [RATA].`,
